webapp_news_aggregation/requirements.txt
#pip install -r requirements.txt

cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
opencv-python==4.1.1.26
pyparsing==2.4.4
python-dateutil==2.8.1
six==1.13.0
torch==1.3.0
wordcloud == 1.6.0
requests == 2.22.0
#feedparser / guizero / newspaper3k / beautifulsoup4
nltk == 3.4.5
gensim == 3.8.1
sklearn

alembic==0.9.6
Babel==2.5.1
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dominate==2.3.1
elasticsearch==7.5.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess_language-spirit==0.5.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
PyJWT==1.5.3
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.2
redis==3.2.1
requests==2.18.4
rq==1.0
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1

